+++ include/bio/ranges/views/char_strictly_to.hpp
@@ -15,9 +15,11 @@
 
 #include <ranges>
 
-#include <bio/alphabet/concept.hpp>
-#include <bio/ranges/views/char_to.hpp>
-#include <bio/ranges/views/validate_char_for.hpp>
+#include <bio/alphabet/custom/all.hpp>
+#include <bio/meta/overloaded.hpp>
+#include <bio/ranges/type_traits.hpp>
+#include <bio/ranges/views/deep.hpp>
+#include <bio/ranges/views/type_reduce.hpp>
 
 namespace bio::ranges::views
 {
@@ -52,13 +54,13 @@ namespace bio::ranges::views
  * | std::ranges::forward_range        |                                     | *preserved*                    |
  * | std::ranges::bidirectional_range  |                                     | *preserved*                    |
  * | std::ranges::random_access_range  |                                     | *preserved*                    |
- * | std::ranges::contiguous_range     |                                     | *lost*                         |
+ * | std::ranges::contiguous_range     |                                     | *lost*¹                        |
  * |                                   |                                     |                                |
  * | std::ranges::viewable_range       | *required*                          | *guaranteed*                   |
  * | std::ranges::view                 |                                     | *guaranteed*                   |
  * | std::ranges::sized_range          |                                     | *preserved*                    |
  * | std::ranges::common_range         |                                     | *preserved*                    |
- * | std::ranges::output_range         |                                     | *lost*                         |
+ * | std::ranges::output_range         |                                     | *lost*¹                        |
  * | std::semiregular                  |                                     | *preserved*                    |
  * | bio::ranges::const_iterable_range |                                     | *preserved*                    |
  * |                                   |                                     |                                |
@@ -66,14 +68,43 @@ namespace bio::ranges::views
  *
  * See the \link views views submodule documentation \endlink for detailed descriptions of the view properties.
  *
+ * ### Special cases
+ *
+ * **Strings ¹:** If a range is given whose value_type is the same its alphabet type (e.g. std::string), a simple view
+ * to the range is returned and no transformation happens. In that case, both, contiguous-ness and writeability are
+ * preserved.
+ * Thus, there is no overhead when applying this adaptor with alphabet_type == char.
+ *
  * ### Example
  *
  * \include test/snippet/ranges/views/char_strictly_to.cpp
  * \hideinitializer
  */
 template <alphabet::alphabet alphabet_type>
-inline auto const char_strictly_to = views::validate_char_for<alphabet_type> | views::char_to<alphabet_type>;
-
+inline auto const char_strictly_to = detail::adaptor_from_functor{meta::overloaded{
+  // clang-format off
+    []<std::ranges::input_range rng_t>(rng_t && range)
+        requires(std::convertible_to<ranges::range_innermost_value_t<rng_t>, alphabet::char_t<alphabet_type>> &&
+                 std::same_as<ranges::range_innermost_value_t<rng_t>, alphabet_type>)
+    {
+        return std::forward<rng_t>(range) | views::type_reduce; // NOP
+    },
+    []<std::ranges::input_range rng_t>(rng_t && range)
+        requires(std::convertible_to<ranges::range_innermost_value_t<rng_t>, alphabet::char_t<alphabet_type>>)
+    {
+        auto fn = [] (alphabet::char_t<alphabet_type> const in) -> alphabet_type
+        {
+            return alphabet::assign_char_strictly_to(in, alphabet_type{});
+        };
+        return std::forward<rng_t>(range) | deep{std::views::transform(fn)};
+    },
+    []<typename rng_t>(rng_t &&)
+    {
+        static_assert(meta::always_false<rng_t>,
+                      "The type you pass to bio::views::char_strictly_to must be a range of elements convertible "
+                      "to the target alphabet's character type.");
+    }}};
+// clang-format on
 //!\}
 
 //!\brief A shortcut for `decltype(std::string_view{} | bio::views::char_strictly_to<alph_t>)`.