+++ include/bio/ranges/views/char_to.hpp
@@ -13,11 +13,11 @@
 
 #pragma once
 
-#include <ranges>
-
-#include <bio/alphabet/concept.hpp>
-#include <bio/meta/type_traits/basic.hpp>
+#include <bio/alphabet/custom/all.hpp>
+#include <bio/meta/overloaded.hpp>
+#include <bio/ranges/type_traits.hpp>
 #include <bio/ranges/views/deep.hpp>
+#include <bio/ranges/views/type_reduce.hpp>
 
 namespace bio::ranges::views
 {
@@ -43,24 +43,31 @@ namespace bio::ranges::views
  * This view is a **deep view** Given a range-of-range as input (as opposed to just a range), it will apply
  * the transformation on the innermost range (instead of the outermost range).
  *
- * | Concepts and traits              | `urng_t` (underlying range type)      | `rrng_t` (returned range type)                     |
- * |----------------------------------|:-------------------------------------:|:--------------------------------------------------:|
- * | std::ranges::input_range         | *required*                            | *preserved*                                        |
- * | std::ranges::forward_range       |                                       | *preserved*                                        |
- * | std::ranges::bidirectional_range |                                       | *preserved*                                        |
- * | std::ranges::random_access_range |                                       | *preserved*                                        |
- * | std::ranges::contiguous_range    |                                       | *lost*                                             |
- * |                                  |                                       |                                                    |
- * | std::ranges::viewable_range      | *required*                            | *guaranteed*                                       |
- * | std::ranges::view                |                                       | *guaranteed*                                       |
- * | std::ranges::sized_range         |                                       | *preserved*                                        |
- * | std::ranges::common_range        |                                       | *preserved*                                        |
- * | std::ranges::output_range        |                                       | *lost*                                             |
- * | bio::ranges::const_iterable_range     |                                       | *preserved*                                        |
- * |                                  |                                       |                                                    |
- * | std::ranges::range_reference_t   | bio::alphabet::char_t<alphabet_t>   | `alphabet_t`                                       |
+ * | Concepts and traits              | `urng_t` (underlying range type)  | `rrng_t` (returned range type)   |
+ * |----------------------------------|:---------------------------------:|:--------------------------------:|
+ * | std::ranges::input_range         | *required*                        | *preserved*                      |
+ * | std::ranges::forward_range       |                                   | *preserved*                      |
+ * | std::ranges::bidirectional_range |                                   | *preserved*                      |
+ * | std::ranges::random_access_range |                                   | *preserved*                      |
+ * | std::ranges::contiguous_range    |                                   | *lost*¹                          |
+ * |                                  |                                   |                                  |
+ * | std::ranges::viewable_range      | *required*                        | *guaranteed*                     |
+ * | std::ranges::view                |                                   | *guaranteed*                     |
+ * | std::ranges::sized_range         |                                   | *preserved*                      |
+ * | std::ranges::common_range        |                                   | *preserved*                      |
+ * | std::ranges::output_range        |                                   | *lost*¹                          |
+ * | bio::ranges::const_iterable_range|                                   | *preserved*                      |
+ * |                                  |                                   |                                  |
+ * | std::ranges::range_reference_t   | bio::alphabet::char_t<alphabet_t> | `alphabet_t`                     |
  *
  * See the \link views views submodule documentation \endlink for detailed descriptions of the view properties.
+ *
+ * ### Special cases
+ *
+ * **Strings ¹:** If a range is given whose value_type is the same its alphabet type (e.g. std::string), a simple view
+ * to the range is returned and no transformation happens. In that case, both, contiguous-ness and writeability are
+ * preserved.
+ * Thus, there is no overhead when applying this adaptor with alphabet_type == char.
  *
  * ### Example
  *
@@ -68,14 +75,30 @@ namespace bio::ranges::views
  * \hideinitializer
  */
 template <alphabet::alphabet alphabet_type>
-inline auto const char_to = deep{std::views::transform(
-  [](auto && in)
-  {
-      static_assert(std::common_reference_with<decltype(in), alphabet::char_t<alphabet_type>>,
-                    "The innermost value type must have a common reference to underlying char type of alphabet_type.");
-      // call element-wise assign_char from the alphabet
-      return alphabet::assign_char_to(in, alphabet_type{});
-  })};
+inline auto const char_to = detail::adaptor_from_functor{meta::overloaded{
+  // clang-format off
+    []<std::ranges::input_range rng_t>(rng_t && range)
+        requires(std::convertible_to<ranges::range_innermost_value_t<rng_t>, alphabet::char_t<alphabet_type>> &&
+                 std::same_as<ranges::range_innermost_value_t<rng_t>, alphabet_type>)
+    {
+        return std::forward<rng_t>(range) | views::type_reduce; // NOP
+    },
+    []<std::ranges::input_range rng_t>(rng_t && range)
+        requires(std::convertible_to<ranges::range_innermost_value_t<rng_t>, alphabet::char_t<alphabet_type>>)
+    {
+        auto fn = [] (alphabet::char_t<alphabet_type> const in) -> alphabet_type
+        {
+            return alphabet::assign_char_to(in, alphabet_type{});
+        };
+        return std::forward<rng_t>(range) | deep{std::views::transform(fn)};
+    },
+    []<typename rng_t>(rng_t &&)
+    {
+        static_assert(meta::always_false<rng_t>,
+                      "The type you pass to bio::views::char_strictly_to must be a range of elements convertible "
+                      "to the target alphabet's character type.");
+    }}};
+// clang-format on
 
 //!\}