+++ test/unit/ranges/views/view_to_char_test.cpp
@@ -37,6 +37,16 @@ TEST(view_to_char, basic)
     EXPECT_EQ(cmp2, v3);
 }
 
+TEST(view_to_char, preserve_string)
+{
+    std::string const s{"ACTTTGATA"};
+
+    auto v = s | bio::ranges::views::to_char;
+
+    EXPECT_EQ(v, s);
+    EXPECT_TRUE((std::same_as<decltype(v), std::string_view>));
+}
+
 TEST(view_to_char, concepts)
 {
     std::vector<bio::alphabet::dna5> vec{"ACTTTGATA"_dna5};
@@ -44,6 +54,7 @@ TEST(view_to_char, concepts)
     EXPECT_TRUE(std::ranges::forward_range<decltype(vec)>);
     EXPECT_TRUE(std::ranges::bidirectional_range<decltype(vec)>);
     EXPECT_TRUE(std::ranges::random_access_range<decltype(vec)>);
+    EXPECT_TRUE(std::ranges::contiguous_range<decltype(vec)>);
     EXPECT_FALSE(std::ranges::view<decltype(vec)>);
     EXPECT_TRUE(std::ranges::sized_range<decltype(vec)>);
     EXPECT_TRUE(std::ranges::common_range<decltype(vec)>);
@@ -55,10 +66,24 @@ TEST(view_to_char, concepts)
     EXPECT_TRUE(std::ranges::forward_range<decltype(v1)>);
     EXPECT_TRUE(std::ranges::bidirectional_range<decltype(v1)>);
     EXPECT_TRUE(std::ranges::random_access_range<decltype(v1)>);
+    EXPECT_FALSE(std::ranges::contiguous_range<decltype(v1)>);
     EXPECT_TRUE(std::ranges::view<decltype(v1)>);
     EXPECT_TRUE(std::ranges::sized_range<decltype(v1)>);
     EXPECT_TRUE(std::ranges::common_range<decltype(v1)>);
     EXPECT_TRUE(bio::ranges::const_iterable_range<decltype(v1)>);
     EXPECT_FALSE((std::ranges::output_range<decltype(v1), bio::alphabet::dna5>));
     EXPECT_FALSE((std::ranges::output_range<decltype(v1), char>));
+
+    std::string s;
+    auto        v2 = s | bio::ranges::views::to_char;
+    EXPECT_TRUE(std::ranges::input_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::forward_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::bidirectional_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::random_access_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::contiguous_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::view<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::sized_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::common_range<decltype(v2)>);
+    EXPECT_TRUE(bio::ranges::const_iterable_range<decltype(v2)>);
+    EXPECT_TRUE((std::ranges::output_range<decltype(v2), char>));
 }